// frontend/src/api-generated/models/hangSanXuatResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface HangSanXuatResponse {
  id?: number;
  maHang?: string;
  tenHang?: string;
  websiteHoTro?: string;
  hotlineHoTro?: string;
  emailHoTro?: string;
  ghiChu?: string;
  trangThai?: string;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
}