// frontend/src/api-generated/models/hangSanXuatRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface HangSanXuatRequest {
  /**
     * @minLength 0
     * @maxLength 50
     */
  maHang?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  tenHang: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  websiteHoTro?: string;
  /**
     * @minLength 0
     * @maxLength 20
     */
  hotlineHoTro?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  emailHoTro?: string;
  ghiChu?: string;
  /**
     * @minLength 0
     * @maxLength 30
     */
  trangThai: string;
}